Added 'rotate_vars' to props at load-time.
This commit standardizes the way rotations are handled. The
'rotate_vars' property is now a set, and we always use it to as the list
of variables to rotate. Also, this includes a fix so that the
orientation matrix always represents the instrument orientation in the
current reference frame (as long as the adv.rotate functions are used).