This fixes a read bug.
I believe the issue here is that the ADV spits out it's own timestamp in the middle of a read period (presumably when its internal clock advances). Here I've ignored those, and just taken the time stamp from the start of the burst. I still need to decide if I want to do something with this timestamp though, or if this introudces other errors?